refactor(experimental): moved deployment module back to experimental
