+++ examples/package.json
@@ -9,7 +9,7 @@
     "hello-advanced": "tsx basic/hello-world.ts",
     "hello-advanced-pool": "tsx pool/hello-world.ts",
     "local-image": "tsx local-image/serveLocalGvmi.ts",
-    "deployment": "tsx deployment/new-api.ts",
+    "deployment": "tsx experimental/deployment/new-api.ts",
     "market-scan": "tsx market/scan.ts",
     "experimental-server": "tsx experimental/express/server.ts",
     "lint": "npm run lint:ts",
