+++ src/experimental/deployment/deployment.ts
@@ -1,15 +1,16 @@
-import { GolemAbortError, GolemUserError } from "../shared/error/golem-error";
-import { defaultLogger, Logger, YagnaApi } from "../shared/utils";
+import { GolemAbortError, GolemUserError } from "../../shared/error/golem-error";
+import { defaultLogger, Logger, YagnaApi } from "../../shared/utils";
 import { EventEmitter } from "eventemitter3";
-import { ActivityModule } from "../activity";
-import { Network, NetworkOptions } from "../network";
-import { GftpStorageProvider, StorageProvider, WebSocketBrowserStorageProvider } from "../shared/storage";
+import { ActivityModule } from "../../activity";
+import { Network, NetworkOptions } from "../../network";
+import { GftpStorageProvider, StorageProvider, WebSocketBrowserStorageProvider } from "../../shared/storage";
 import { validateDeployment } from "./validate-deployment";
-import { DemandBuildParams, DraftOfferProposalPool, MarketModule } from "../market";
-import { PaymentModule } from "../payment";
+import { DemandBuildParams, DraftOfferProposalPool, MarketModule } from "../../market";
+import { PaymentModule } from "../../payment";
 import { CreateActivityPoolOptions } from "./builder";
 import { Subscription } from "rxjs";
-import { LeaseProcessPool, LeaseProcessPoolOptions } from "../lease-process";
+import { LeaseProcessPool, LeaseProcessPoolOptions } from "../../lease-process";
+import { DataTransferProtocol } from "../../shared/types";
 
 export enum DeploymentState {
   INITIAL = "INITIAL",
@@ -47,8 +48,6 @@ export type DeploymentComponents = {
   activityPools: { name: string; options: CreateActivityPoolOptions }[];
   networks: { name: string; options: NetworkOptions }[];
 };
-
-export type DataTransferProtocol = "gftp" | "ws" | StorageProvider;
 
 export interface DeploymentOptions {
   dataTransferProtocol?: DataTransferProtocol;
