+++ src/experimental/index.ts
@@ -1,2 +1,3 @@
 export * from "./job";
 export * from "./reputation";
+export * from "./deployment";
