+++ src/index.ts
@@ -1,14 +1,12 @@
 //  High-level entry points
 export * from "./golem-network";
-export * from "./deployment";
+export * from "./lease-process";
 
 // Low level entry points for advanced users
 export * from "./market";
 export * from "./payment";
 export * from "./network";
 export * from "./activity";
-export * from "./market/agreement";
-export * from "./lease-process";
 
 // Necessary domain entities for users to consume
 export * from "./shared/error/golem-error";
