+++ src/shared/types.ts
@@ -1,0 +1,3 @@
+import { StorageProvider } from "./storage";
+
+export type DataTransferProtocol = "gftp" | "ws" | StorageProvider;
